requirements.txt
asgiref==3.8.1
Django==5.1.1
django-pytest==0.2.0
djangorestframework==3.15.2
iniconfig==2.0.0
mysqlclient==2.2.4
packaging==24.1
pika==1.3.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.3.3
sqlparse==0.5.1
